requirements.txt
langchain
langchain_community
langchain-google-genai
python_dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
unstructured